[Arduino IDE] Fix compile issues when selecting specific controllers
