+++ src/_C001.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C001
 //#######################################################################################################
 //########################### Controller Plugin 001: Domoticz HTTP ######################################
