+++ src/_C002.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C002
 
 // #######################################################################################################
