+++ src/_C003.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C003
 //#######################################################################################################
 //########################### Controller Plugin 003: Nodo Telnet  #######################################
