+++ src/_C005.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C005
 //#######################################################################################################
 //################### Controller Plugin 005: Home Assistant (openHAB) MQTT ##############################
