+++ src/_C007.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C007
 //#######################################################################################################
 //########################### Controller Plugin 007: Emoncms ############################################
