+++ src/_C008.ino
@@ -1,3 +1,5 @@
+#include "_CPlugin_Helper.h"
+
 #ifdef USES_C008
 //#######################################################################################################
 //########################### Controller Plugin 008: Generic HTTP #######################################
