+++ src/_C009.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C009
 //#######################################################################################################
 //########################### Controller Plugin 009: FHEM HTTP ##########################################
