+++ src/_C010.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C010
 //#######################################################################################################
 //########################### Controller Plugin 010: Generic UDP ########################################
