+++ src/_C011.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C011
 //#######################################################################################################
 //########################### Controller Plugin 011: Generic HTTP Advanced ##############################
