+++ src/_C012.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C012
 //#######################################################################################################
 //########################### Controller Plugin 012: Blynk  #############################################
