+++ src/_C013.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C013
 
 #include "src/Globals/Nodes.h"
