+++ src/_C014.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C014
 
 #include "src/Globals/Device.h"
