+++ src/_C016.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C016
 //#######################################################################################################
 //########################### Controller Plugin 016: Controller - Cache #################################
