+++ src/_C017.ino
@@ -1,3 +1,4 @@
+#include "_CPlugin_Helper.h"
 #ifdef USES_C017
 //#######################################################################################################
 //###########################   Controller Plugin 017: ZABBIX  ##########################################
